+++ README.md
@@ -11,6 +11,7 @@ Welcome to the BudgetChain Frontend! This project is built using Next.js and ser
 - [Contributing](#contributing)
 - [License](#license)
 - [Contact](#contact)
+- [Settings Page Implementation](#settings-page-implementation)
 
 ## Features
 - **User Authentication**: Secure login and registration for users.
@@ -93,3 +94,76 @@ For any questions or inquiries, please reach out to the project maintainers:
 - **GitHub** - [anonfedora](https://github.com/anonfedora)
 
 Thank you for your interest in BudgetChain! We hope you find this project useful and enjoyable to work with. Happy coding!
+
+## Settings Page Implementation
+
+### Overview
+Implemented a new settings page for the dashboard that allows users to manage their preferences and notification settings. The page follows the existing dashboard design language and provides a seamless user experience.
+
+### Features
+
+#### User Interface
+- Modern, responsive design using Shadcn UI components
+- Dark theme consistent with the dashboard
+- Tab-based navigation for different settings sections
+- Form validation and error handling
+- Loading states and feedback notifications
+
+#### Functionality
+- **Preferences Management**
+  - Language selection
+  - Timezone configuration
+  - Theme preferences
+- **Notification Settings**
+  - Email notifications toggle
+  - Push notifications toggle
+  - Project updates notifications toggle
+
+#### Technical Implementation
+- Built with React and Next.js
+- Form handling using react-hook-form
+- Validation using Zod schema
+- State persistence using localStorage
+- Toast notifications using react-hot-toast
+- Responsive design with Tailwind CSS
+
+### Technical Details
+
+#### Form Validation
+```typescript
+const settingsSchema = z.object({
+  language: z.string().min(1, "Language is required"),
+  timezone: z.string().min(1, "Timezone is required"),
+  theme: z.string().min(1, "Theme is required"),
+  emailNotifications: z.boolean(),
+  pushNotifications: z.boolean(),
+  projectUpdates: z.boolean(),
+});
+```
+
+#### State Management
+- Form state managed by react-hook-form
+- Settings persisted in localStorage
+- Real-time validation feedback
+- Loading states for async operations
+
+#### UI Components
+- Tabs for section navigation
+- Cards for content organization
+- Input fields with validation
+- Checkboxes for toggle options
+- Buttons with loading states
+
+### Accessibility
+- Semantic HTML structure
+- Proper ARIA labels
+- Keyboard navigation support
+- Color contrast compliance
+- Responsive design for all screen sizes
+
+### Future Improvements
+- Integration with backend API
+- Additional settings sections
+- Theme customization options
+- User profile management
+- Advanced notification settings
